pyproject.toml
[tool.poetry]
name = "campaignReport-tap-sendinblue"
version = "0.0.1"
description = "`tap-sendinblue` is a Singer tap for SendInBlue, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Sistemas Interno"]
keywords = [
    "ELT",
    "SendInBlue",
]
license = "Apache 2.0"
packages = [
    { include = "tap_sendinblue" },
]

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.22.0" }
fs-s3fs = { version = "^1.1.1", optional = true }
requests = "^2.28.2"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "^5.0.4"
darglint = "^1.8.1"
black = "^23.1.0"
pyupgrade = "^3.3.1"
mypy = "^1.0.0"
isort = "^5.11.5"
python-decouple = "^3.8"
singer-sdk = { version="^0.22.0", extras = ["testing"] }
types-requests = "^2.28.11.12"

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_sendinblue"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-sendinblue = 'tap_sendinblue.tap:TapSendInBlue.cli'
